Added new fields to the Course Class
