+++ src/main/java/com/archisacadeny/instructor/Instructor.java
@@ -4,49 +4,5 @@ import java.awt.*;
 
 public class Instructor {
 
-    private long id;
-    private String fullName;
-    private String gender;
-    private String department;
-    private List courses;
 
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public String getDepartment() {
-        return department;
-    }
-
-    public void setDepartment(String department) {
-        this.department = department;
-    }
-
-    public List getCourses() {
-        return courses;
-    }
-
-    public void setCourses(List courses) {
-        this.courses = courses;
-    }
 }
